requirements.txt
altair
pandas
streamlit
plotly.express
mathModel
numpy